CB-11968: Add support for config-file in config.xml
